* if it's used as a double, make it a double ;)
git-svn-id: svn+ssh://c4science.ch/source/amie/trunk@3384 64fb8e13-074e-4207-b257-2d34fb7690c6

+++ features/features.cpp
@@ -1470,7 +1470,7 @@ void FeatureTree::stitch()
 
 }
 
-void FeatureTree::setSamplingNumber ( size_t news )
+void FeatureTree::setSamplingNumber ( double news )
 {
     samplingNumber = news ;
     needMeshing = true ;
